% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom.contact.r
\name{bottom.contact}
\alias{bottom.contact}
\title{bottom.contact}
\usage{
bottom.contact(x, bcp, debugrun = FALSE)
}
\value{
dataframe with various results
}
\description{
Calculates indicies of touchdown and liftoff
}
